fix: diffusion head 回退到 float16 (与 LLM 一致防止混合精度梯度错误)

+++ src/vla/model_loader.py
@@ -148,7 +148,7 @@ class OpenVLAPolicy:
 
         self.diffusion_head = DiffusionActionHead(
             hidden_dim=hidden_dim, chunk_size=self.chunk_size,
-        ).to(device=self.device, dtype=torch.float32)
+        ).to(device=self.device, dtype=self.llm.dtype)
 
         import torchvision.transforms as T
         self.img_preprocess = T.Compose([
